pyproject.toml
[tool.poetry]
name = "binanceBot"
version = "0.1.0"
description = "binance data colloctor"
authors = ["cinuor <cokie@foxmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-binance = "^1.0.15"
fire = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
